Reorder article folders and split some files
