+++ src/clients/artikkelit/actions/articleReset.js
@@ -1,5 +1,5 @@
 import {showSnackbar} from '/common/ui-utils.js';
-import {idbClear, getTableNames} from '/artikkelit/indexDB.js';
+import {idbClear, getTableNames} from '/artikkelit/utils/indexDB.js';
 import {resetAndHideCcLicense} from '/artikkelit/interfaces/article.js';
 import {resetPublicationSearchResultSelect} from '/artikkelit/interfaces/publicationSearch.js';
 import {resetReviewSearchResultSelect} from '/artikkelit/interfaces/reviewSearch.js';
