+++ src/clients/artikkelit/actions/articleUpdate.js
@@ -1,6 +1,6 @@
 import {generateArticleRecord} from '/common/rest.js';
 import {showRecord} from '/common/marc-record-ui.js';
-import {idbClear, idbGet, idbSet, idbGetStoredValues} from '/artikkelit/indexDB.js';
+import {idbClear, idbGet, idbSet, idbGetStoredValues} from '/artikkelit/utils/indexDB.js';
 import {refreshAbstractList} from '/artikkelit/interfaces/abstracts.js';
 import {refreshNotesList, refreshOtherTitlesList, refreshUDKsList, refreshOtherRatingsList} from '/artikkelit/interfaces/additionalFields.js';
 import {refreshSciencesList, refreshMetodologysList} from '/artikkelit/interfaces/sciencesAndMethodologies.js';
@@ -8,7 +8,7 @@ import {refreshAuthorsList, refreshAuthorOrganizationList} from '/artikkelit/int
 import {journalTemplate, bookTemplate} from '/artikkelit/constants/index.js';
 import {refreshOntologyWordList} from '/artikkelit/interfaces/ontologyWords.js';
 import {refreshReviewsList} from '/artikkelit/interfaces/reviewSearch.js';
-import {resetCheckAndSave} from '/artikkelit/actions/articleActions.js';
+import {resetCheckAndSave} from '/artikkelit/actions/articleModes.js';
 
 
 export function collectFormData() {
