+++ src/clients/artikkelit/interfaces/ontologyWords.js
@@ -1,5 +1,5 @@
-import {idbAddValueToLastIndex, idbClear, idbDel, idbGetStoredValues} from '/artikkelit/indexDB.js';
-import {addValueToSessionStoreList, getSessionStoreValue, resetSessionStoreList} from '/artikkelit/sessionStorageManager.js';
+import {idbAddValueToLastIndex, idbClear, idbDel, idbGetStoredValues} from '/artikkelit/utils/indexDB.js';
+import {addValueToSessionStoreList, getSessionStoreValue, resetSessionStoreList} from '/artikkelit/utils/sessionStorageManager.js';
 import {getOntologyWords} from '/common/rest.js';
 import {formToJson, createIconButton, createP, setOptions, showSnackbar} from '/common/ui-utils.js';
