+++ src/clients/artikkelit/interfaces/sciencesAndMethodologies.js
@@ -1,4 +1,4 @@
-import {idbAddValueToLastIndex, idbClear, idbDel, idbGetStoredValues, } from '/artikkelit/indexDB.js';
+import {idbAddValueToLastIndex, idbClear, idbDel, idbGetStoredValues, } from '/artikkelit/utils/indexDB.js';
 import {createIconButton, createP, formToJson, showSnackbar} from '/common/ui-utils.js';
